split transcribe.js as sphinx.js and kaldi.js, fix create socket function
